Big refactor to a lib.rs.
